+++ lib/src/collections.dart
@@ -4,6 +4,7 @@ import 'dart:math' as dart_math;
 import 'date.dart';
 import 'math.dart';
 import 'utils.dart';
+import 'weak_map.dart';
 
 /// Represents a pair with [a] and [b] of type [T].
 class Pair<T> {
@@ -2496,12 +2497,17 @@ class ObjectCache {
 ///
 /// Browser: one useful way is to use with [document] (the root of DOM),
 /// and be able to associate values with any [Node] in DOM tree.
-class TreeReferenceMap<K, V> implements Map<K, V> {
+class TreeReferenceMap<K extends Object, V extends Object>
+    implements Map<K, V> {
   /// The root of the Tree Reference.
   final K root;
 
   /// If true, each operation performs a purge.
   final bool autoPurge;
+
+  static const int defaultAutoPurgeThreshold = 100;
+
+  final int autoPurgeThreshold;
 
   /// Will stored purged entries in a separated [Map].
   final bool keepPurgedEntries;
@@ -2521,17 +2527,21 @@ class TreeReferenceMap<K, V> implements Map<K, V> {
   /// The [Function] that returns true if [parent] has [child].
   final bool Function(K parent, K child, bool deep)? childChecker;
 
+  void Function(Map<K, V> purgedEntries)? onPurgedEntries;
+
   TreeReferenceMap(this.root,
       {this.autoPurge = false,
+      this.autoPurgeThreshold = defaultAutoPurgeThreshold,
       bool keepPurgedKeys = false,
       this.purgedEntriesTimeout,
       this.maxPurgedEntries,
       this.parentGetter,
       this.childrenGetter,
-      this.childChecker})
+      this.childChecker,
+      this.onPurgedEntries})
       : keepPurgedEntries = keepPurgedKeys;
 
-  final Map<K, V> _map = {};
+  final WeakKeyMap<K, V> _map = WeakKeyMap(autoPurge: false);
 
   void put(K key, V value) {
     _map[key] = value;
@@ -2689,7 +2699,7 @@ class TreeReferenceMap<K, V> implements Map<K, V> {
   bool isChildOf(K? parent, K? child, bool deep) =>
       parent != null && child != null && childChecker!(parent, child, deep);
 
-  Map<K, MapEntry<DateTime, V>>? _purged;
+  DualWeakMap<K, MapEntry<DateTime, V>>? _purged;
 
   /// Returns the purged entries length. Only relevant if [keepPurgedEntries] is true.
   int get purgedLength => _purged != null ? _purged!.length : 0;
@@ -2708,9 +2718,18 @@ class TreeReferenceMap<K, V> implements Map<K, V> {
 
   int get purgedEntriesCount => _purgedEntriesCount;
 
+  int _unpurgedCount = 0;
+
   /// Remove all [invalidKeys].
   TreeReferenceMap<K, V> purge() {
+    _unpurgedCount = 0;
+
+    final onPurgedEntries = this.onPurgedEntries;
+
     var changed = false;
+
+    var purgedEntries = <K, V>{};
+
     if (keepPurgedEntries) {
       revalidatePurgedEntries();
       checkPurgedEntriesTimeout();
@@ -2718,7 +2737,7 @@ class TreeReferenceMap<K, V> implements Map<K, V> {
       var invalidKeys = this.invalidKeys;
       if (invalidKeys.isEmpty) return this;
 
-      var purged = _purged ?? <K, MapEntry<DateTime, V>>{};
+      var purged = _purged ?? DualWeakMap(autoPurge: false);
       _purged = purged;
 
       for (var k in invalidKeys) {
@@ -2727,29 +2746,44 @@ class TreeReferenceMap<K, V> implements Map<K, V> {
         changed = true;
         if (val != null) {
           purged[k] = MapEntry(DateTime.now(), val);
+          if (onPurgedEntries != null) {
+            purgedEntries[k] = val;
+          }
         }
       }
 
       checkPurgeEntriesLimit();
     } else {
       for (var k in invalidKeys) {
-        _map.remove(k);
+        var val = _map.remove(k);
         _purgedEntriesCount++;
         changed = true;
+        if (val != null && onPurgedEntries != null) {
+          purgedEntries[k] = val;
+        }
       }
     }
 
     if (changed) {
       _expireCache();
+
+      if (onPurgedEntries != null) {
+        onPurgedEntries(purgedEntries);
+      }
     }
 
     return this;
   }
 
   /// Same as [purge], but called automatically by many operations.
-  void doAutoPurge() {
-    if (!autoPurge) return;
+  bool doAutoPurge() {
+    if (!autoPurge) return false;
+    if (_unpurgedCount < autoPurgeThreshold) {
+      ++_unpurgedCount;
+      return false;
+    }
     purge();
+    return true;
   }
 
   /// Removed purged entries over [maxPurgedEntries] limit.
@@ -2786,6 +2820,7 @@ class TreeReferenceMap<K, V> implements Map<K, V> {
         purgedEntriesTimeout.inMilliseconds > 0) {
       var timeoutMs = purgedEntriesTimeout.inMilliseconds;
       var now = DateTime.now().millisecondsSinceEpoch;
+
       var expired = purged.entries
           .where((e) => (now - e.value.key.millisecondsSinceEpoch) > timeoutMs)
           .map((e) => e.key)
@@ -2824,6 +2859,7 @@ class TreeReferenceMap<K, V> implements Map<K, V> {
 
     if (revalidateCount > 0) {
       _expireCache();
+      ++_unpurgedCount;
     }
 
     _revalidatedPurgedEntriesCount += revalidateCount;
@@ -2924,6 +2960,7 @@ class TreeReferenceMap<K, V> implements Map<K, V> {
     var rm = _map.remove(key);
     doAutoPurge();
     _expireCache();
+
     return rm;
   }
 
@@ -2937,6 +2974,7 @@ class TreeReferenceMap<K, V> implements Map<K, V> {
   void clear() {
     _map.clear();
     _expireCache();
+    _unpurgedCount = 0;
   }
 
   List<K>? _keysReversedList;
@@ -3018,6 +3056,7 @@ class TreeReferenceMap<K, V> implements Map<K, V> {
     doAutoPurge();
     var val = _map.putIfAbsent(key, ifAbsent);
     _expireCache();
+
     return val;
   }