+++ lib/swiss_knife.dart
@@ -13,3 +13,4 @@ export 'src/resource.dart';
 export 'src/string.dart';
 export 'src/uri.dart';
 export 'src/utils.dart';
+export 'src/weak_map.dart';